+++ root/root_test.go
@@ -1,0 +1,15 @@
+package root
+
+import "testing"
+
+func TestRoot(t *testing.T) {
+	// TODO: NewCommand
+	// TODO: NewCommandMap
+	// TODO: MainLoop
+	// TODO: ParamsMap
+	// TODO: Authenticate
+	// TODO: validateCommandParams
+	// TODO: getCommandParams
+	// TODO: displayHelp
+	// TODO: searchCommand
+}