+++ CHANGELOG.md
@@ -1,12 +1,18 @@
+## v0.3.8 - Root Package
+  * Commit: 2026-04-07 15:13
+  * New Commands
+  * ParamsMap
+  * Get command params
+  * Display help and search command
+  * MainLoop
 ## v0.3.7 - Application Packages 
-  * Commit: 
+  * Commit: 2026-04-07 13:50
   * Fail functions
   * Conf package
   * conf.LoadEnv
   * dict.Inspect
   * Sys package and display functions
   * String JSON functions
-  * Root package
 ## v0.3.6 - Schema Combos 
   * Commit: 2026-04-01 22:44
   * Schema GetOrCreate